+++ libsolidity/analysis/TypeChecker.cpp
@@ -1760,6 +1760,33 @@ bool TypeChecker::visit(FunctionCall const& _functionCall)
 		}
 	}
 
+	if (functionType->takesSinglePackedBytesParameter())
+	{
+		string generalMessage =
+			"This function only accepts a single \"bytes\" argument. Please use "
+			"\"abi.encodePacked(...)\" or a similar function to encode the data.";
+
+		if (arguments.size() > 1)
+		{
+			if (v050)
+				m_errorReporter.typeError(_functionCall.location(), generalMessage);
+			else
+				m_errorReporter.warning(_functionCall.location(), generalMessage);
+		}
+		else if (arguments.size() == 1 && !type(*arguments.front())->isImplicitlyConvertibleTo(ArrayType(DataLocation::Memory)))
+		{
+			string msg =
+				generalMessage +
+				" The provided argument of type " +
+				type(*arguments.front())->toString() +
+				" is not implicitly convertible to expected type bytes memory.";
+			if (v050)
+				m_errorReporter.typeError(_functionCall.location(), msg);
+			else
+				m_errorReporter.warning(_functionCall.location(), msg);
+		}
+	}
+
 	if (functionType->takesArbitraryParameters() && arguments.size() < parameterTypes.size())
 	{
 		solAssert(_functionCall.annotation().kind == FunctionCallKind::FunctionCall, "");